+++ lib/restart-trigger.ts
@@ -1,2 +1,2 @@
-// restart triggered at 2026-03-08T17:54:44.049Z
-export const RESTART_TS = 1772992484049;
+// restart triggered at 2026-03-11T10:09:56.390Z
+export const RESTART_TS = 1773223796390;